v1.4.34: Fix server/browser build regression from ComfyUI version check
