+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://raw.githubusercontent.com/nicerlab/files/main/tauri/tauri.conf.json.schema",
   "productName": "MooshieUI",
-  "version": "1.4.33",
+  "version": "1.4.34",
   "identifier": "com.mooshieui.desktop",
   "build": {
     "frontendDist": "../dist",
